services imlements

+++ src/main/java/com/prueba/model/Cliente.java
@@ -20,7 +20,7 @@ public class Cliente implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int idCliente;
+	private Long idCliente;
 	
 	@Column(name = "nombre")
 	private String nombre;
@@ -46,14 +46,20 @@ public class Cliente implements Serializable{
 	}
 
 
-	public int getIdCliente() {
+	
+
+	public Long getIdCliente() {
 		return idCliente;
 	}
 
 
-	public void setIdCliente(int idCliente) {
+
+
+	public void setIdCliente(Long idCliente) {
 		this.idCliente = idCliente;
 	}
+
+
 
 
 	public String getNombre() {
